refactor: move to input directory Signed-off-by: Philippe Martin <phmartin@redhat.com>
