+++ packages/ui/src/lib/index.ts
@@ -20,8 +20,8 @@ import type { ButtonType } from './button/Button';
 import Button from './button/Button.svelte';
 import Checkbox from './checkbox/Checkbox.svelte';
 import Input from './input/Input.svelte';
+import SearchInput from './input/SearchInput.svelte';
 import Modal from './modal/Modal.svelte';
-import SearchInput from './searchInput/SearchInput.svelte';
 import Spinner from './spinner/Spinner.svelte';
 import Tab from './tab/Tab.svelte';
 import TableDurationColumn from './table/DurationColumn.svelte';
