+++ packages/ui/src/lib/input/SearchInput.svelte
